Remove filter size limit in add_glare
In extreme cases (e.g. a very bright light source in black space),
glare from a single point can visibly affect the entire image, and a
hard limit on the filter size has an obvious and unpleasant
appearance.  So just get rid of the limit.

At the same time, use symmetry to cut the filter-calculation time in
half -- so even with the bigger filter, this version is faster than
the previous one.