Fix use of filter symmetry in add_glare
The previous attempt to use filter x-y symmetry to speed up
calculation of the filter matrix didn't work correctly for non-square
images.

This change fixes that problem, and also uses vertical and horizontal
symmetry to further reduce the amount of calculation done.

A somewhat subtle change is that the filter is now truly symmetric,
rather than being shifted by 0.5 pixels in x and y directions.

So the value at location 0,0 in the filter matrix represents only the
positive quadrant, with the filter center at the origin (whereas
previously it represented all four quadrants around the filter origin,
with half the extent in each direction).