Use simpler implementation for Lua "use" function
This version is basically just like the "include" function -- it just
copies all symbols from the loaded file into the current environment
-- but avoids using the same file multiple times

The old implementation of "use" was sort of clever, and maybe more
efficient, but doesn't work properly if use'd functions reference
things that are only defined in the target environment.  Since the
"scene" and "camera" variables are now defined that way, this caused
some scene files to break.

The new implementation is simple, stupid, but straight-forward, and
works properly for the above case.

The "use" function is kind of ugly anyway, really scene files should
use a more structured module mechanism of some sort, like Lua's
require (maybe augmented a little for snogray-specific niceties)...