Change the way parameter "sub-options" are handled
This change affects command-line parsing via
CmdLineParser::store_opt_arg_with_sub_options, and the various places
the resulting parameters are read.

Instead of storing a "main" parameter as NAME, and the associated
sub-parameters as NAME.MAIN_VAL.SUB_PARAM_KEY, store the main
parameter as NAME.MAIN_PARAM_KEY and the sub-parameters as
NAME.SUB_PARAM_KEY.

This is a little simpler to handle generally, and not having a string
parameter with the same name as a prefix will be necessary when we
move to real sub-tables later.