Fix snogcvt command-line handling to require image args
Despite the assertion in the help output, snogcvt couldn't actually
use standard-input/output for image i/o when no image filename was
specified on the command-line; the argument parsing would allow it,
but it would just crash later on.

Since this capability isn't really so useful, and it _is_ useful to
give a usage message when no command-line arguments are provided, just
do that instead.