Replace dashes in parameter names with underscores
Dashes are a bit prettier, but underscores seem a better bet to work
smoothly with Lua tables (Lua allows dashes in table keys of course,
but the nice syntax sugar only works for keys that are valid Lua
identifiers).

To allow users to use the prettier names, though, the command-line
interface now changes dashes in parameter names to underscores.