Rewrite recover_image to open the source image itself
It's a bit annoying, as the partial-image filename may not be a normal
filename, so we need to determine its format using the _output_
filename.

However doing it this way is a little more portable as it doesn't
require renaming an open file: previously it had to open the old
[partially written] output file and then rename it before opening the
new output file with the same name.  Now the renaming can be done first.