Rewrite top-level in Lua
This replaces much of the "top level" ("main" function etc) of
snogray, which used to be written in C++, with one written in Lua.
That includes command-line parsing, setup, the invocation of the
various phases, timing, and final statistics reporting.

Right now this is pretty much an exact duplicate of the old C version,
but hopefully Lua should make it simpler to add features and alternate
top-level drivers.