Use term "radius" instead of "width" for image filters
The terminology used for image filter coverage is a bit confused, and
the term "radius" seems to better match most usage: it's the distance
from sample centers to the edge of the filter.  "Radius" isn't
entirely correct either, as separable filters treat the two image
dimensions independently, but at least it's less confusing.