Use GaussianSqrtFilter in PhotonEval::Lo
The previous filter was directly from Henrik Wann Jensen's
book ("Realistic Image Synthesis Using Photon Mapping").  It wasn't a
straight gaussian, and I'm not really sure what exactly it was doing
or how the parameters used were derived.

This new filter is conventional and easy to understand, and seems to
work equally well.