Rework Lua command-line parsing a bit
Rename the cmdlineparser creation function to "cmdlineparser.parser".

Add a new "cmdlineparser.standard_parser" function that returns a
parser with standard GNU options and behavior already enabled.

cmdlineparser.add_standard_options now modifies the parser passed as an
argument, rather than returning a new parser with additional
options (this is somewhat more convenient in the case where the user
wants to add additional options later).

Support a new "final_options" member in parsers, which is simply
treated as an additional list of options that is processed following
other options in the parser.

For standard parsers, add a "Try 'prog --help' ..."  message to
various errors.