Rewrite module-setup.lua handling of non-installed modules
Instead of capturing all module names starting with "snogray." and
returning a function thatdelays the actual searching until load time,
just try loading the module in the searcher, returning nil when we
can't find anything.  This allows the standard Lua searchers a crack
at finding the module when ours can't.

This is cleaner conceptually (loadfile does exactly what we need,
reading the file contents and returning a function that instantiates
the load) and also allows "snogray." modules to be loaded via other
searchers, most importantly the preload searcher.