Move Light render interface into new Light::Sampler class
This basically splits the old Light class into two parts:  The Light
class continues to be used for scene-definition, and a new
Light::Sampler class is used for rendering.  Most of the complexity of
lights is actually in Light::Sampler.

Light::Sampler objects are created just before rendering, after
loading the entire scene, and so can naturally incorporate any extra
data needed for rendering that can't be calculated until the entire
scene is available (such as used by EnvmapLight and FarLight).
Previously we just delayed initializing that data, and called
Light::setup just before rendering to calculate it, but this forced us
to always keep lights mutable, which was kind of ugly (now we can make
Lights const in a bunch of places we couldn't before).

This will also make it a bit easier to add explicit lights to
Surfaces.  With the old scheme, surfaces just added lights directly to
a global list for rendering.  That was fine for area lights, which
don't exist until that point, but any explicit lights would have
ownership issues:  both the containing Surface and the global list
would think they had ownership.  Light::Sampler objects are always
created for rendering, and so have a clear owner.