Remove container functionality of Scene
This makes Scene a purely render-time abstraction, which references
the scene contents indirectly.

As there is no split into "scene-setup" and "render" states, we can
get rid of the Scene::setup method, etc, and do everything in the
constructor.

For now, Scene::num_surfaces is disabled, as Scene now just contains a
simple Surface reference in Scene::root_surface, and there's no way to
count surfaces.