Make Scene object a part of GlobalRenderState
Now that the Scene object is only used during rendering, there's no
reason for it to be created independently, so just embed it directly
in GlobalRenderState, which allows a somewhat more convenient
interface.

This also moves selection of the search accelerator into the
GlobalRenderState constructor, which is where many of the other
render-time configuration is done.

Printing of scene-statistics is currently disabled.