Optimize Octree search by using ray parametric form
Instead of using positions to search in an Octree, formulate the
search in terms of the ray's parameter space.  As the goal is to find
intersections with the ray, anything of interest can by definition be
expressed this way, and doing so is both more efficient and simpler.

In addition, child-nodes are now searched starting with those closest
to the ray's origin, which in some cases results in faster pruning of
the search space.