Require SWIG "disown" support
Avoiding memory leaks and/or corruption without SWIG "disown" support
is tricky and intrusive, and the existing Snogray code which attempts
to do so is bit-rotten and hard to fix.

SWIG has offered Lua "disown" support for about five years (since
version 1.3.35), so it seems reasonable to just require it.