Fix Octree::for_each_possible_intersector edge cases
When a ray is absolutely aligned with an axis, just treat it as having
an extremely shallow slope instead of using zero for that axis's
parametric boundaries.

The former essentially just yields very large negative/positive
parametric bounds, which work fine (they'll be limited by other bounds
anyway), whereas the latter results in a zero-sized parametric region,
which causes the search to immediately fail.