Don't create null references in Ref
Ref used to get the underlying pointer of another Ref object by doing
"&*OTHER", where the "*" would invoked Ref::operator*.  In the case
where OTHER.obj is null, this would return a "null reference" which is
not a valid thing in C++.

This actually worked correctly in most cases, because we'd immediately
take the address, and given the way C++ compilers implement
references, it would all sort of work out.  However it's still not
valid code, and was causing crashes when compiled with recent versions
of clang.

Fix this simply by directly getting the pointer, instead of going
through a reference.