configure.ac: Disable use of "extern template" by default
Although there's a configure test for the extern-template problem with
gcc 4.9 (which seems to be fixed in the gcc development versio),
another, different, problem with mixing extern-template and -flto has
shown up in the gcc development version.

The gain from using extern template isn't all that huge, so for the
moment it seems simpler just not to use it.